[layout] Use MOV32mr instead of MOV32ri with store and CopyToReg as users.
When the same SD constant is used in X86 by a store and a CopyToReg
as users, we want to reuse the same virtual register for this,
since this is what AArch64 does. Therefore, we try to match with
MOV32mr, so that the constant should come from a register.

Addresses: https://github.com/systems-nuts/UnASL/issues/76