[layout] Use only temporary registers for ADRP instructions (not coming from MOVaddr).
With consecutive calls of the same global argument (by value),
the compiler will use a CSR as the destination of the ADRP.
This register needs to be pushed, but X86 has rip-relative
mem-to-reg moves, which avoid this issue. So, we need to use
temporary registers in the case of AArch64.

Addresses: https://github.com/systems-nuts/UnASL/issues/102