[layout] Add flag for reserving a register scavenging spill slot.
AArch64 tries to reserve an extra CSR during register scavenging,
in order to use it to access stack offsets larger than 255,
which cannot be encoded directly to instructions as an offset.
If the stack size is larger than 255, but AArch64 cannot find a CSR,
it will reserve a special spill slot for that purpose.
To align that behaviour with X86, we add a flag to always (conservatively)
assign this spill slot for both architectures.

Addresses: https://github.com/systems-nuts/UnASL/issues/110