[layout] Make check for simplified addressing with register offset SHL-specific.
We add a more specific check for the case that solves:
042f2f81e548e07c2bf4199c61117d6dfcd5cfcd
From now on, we will check if at least one of the operands of the add
node in the DAG is of SHL type, which probably means that we are
in the case of addressing with scaled-register offset.
Otherwise, the previous solution was introducing assymetries for
other addressing mode cases.

Addresses: https://github.com/systems-nuts/UnASL/issues/129