[layout] Fix alignment when force-adding an emergency spill slot.
If the number of CSRs is odd, then the spill slot can go after the
last CSR, at 16-byte alignment. Otherwise, it should go after the
last CSR at 8-byte alignment. Programs with big stack exhibited the
erroneous behavior, because otherwise the CSStack was combined with
the local stack in the PEI, and the error was hidden.

Addresses: https://github.com/systems-nuts/UnASL/issues/137