[layout] Avoid wide multiplication-addition instructions in AArch64.
AArch64 supports instructions where the source operands are 32-bit,
whereas the destination is 64-bit. In X86, all operands will be 64-bit,
for a 64-bit destination, which leads to differences in register allocation,
in case there is a register used both as 32-bit and as 64-bit:
In X86 only one register will be used for that,
whereas AArch64 will employ one 32-bit and one 64-bit register.

Addresses: https://github.com/systems-nuts/UnASL/issues/179