[layout] AArch64 GPR Register Costs and Allocation Order (#56)
* Undo assigning zero cost per use for r15 and its subregs.

The tests seem to pass anyway, and we will substitute this with a
cleaner way, which is to actually assign the cost per use for the
AArch64 registers.

* Apply the X86 register costs to the reciprocal AArch64 registers

In X86, registers r8 to r15 (and their subregisters) have cost per use of one,
due to encoding reasons.  Apply, this cost to the reciprocal registers
(and subregisters) of AArch64.

* Reorder the definition of GPRs in AArch64 to match the one in X86

X86 registers are defined in a specific order which reflects the cost
per use of each register. To ensure the same allocation decisions are
taken for both architectures, we need to define the same order for the
various GPR classes in the AArch64 register info target description.