Force alignment of all objects in the stack to be at least four bytes (#57)
This happens inside the stack-slot-coloring pass. For some reason, there
are cases where our `-align-bytes-to-four` flag does not work, for
example when lowering instructions that use `ccond`, which uses 32 bits
in AArch64 and 8 bits in X86. We add an extra flag to hard-code that
all stack objects should be aligned at a four-byte boundary at least.

Addresses: https://github.com/systems-nuts/unifico/issues/279