Convert AArch64 ADD/SUB-rr instructions to the two-address format (#59)
* [layout] AArch64: use the two-address instruction format for ADD/SUB

It seems that in many cases, the only way to control the register
allocation is to impose the same two-address instruction format between
the architectures. Otherwise, the different live ranges lead to
different allocation of CSRs. Also, need to make these instructions
commutable, in order to allow other smart tricks of the
`twoaddressinstruction` pass (e.g., to enable more coalescing).

Maybe in the future we'll find a better way.

Addresses: https://github.com/systems-nuts/unifico/issues/283

* [layout] Convert ADD32rr into LEA32r instead of LEA64_32r

During the `twoaddressinstruction` pass, `ADD32rr` instructions
are converted into LEA instructions using the `convertToThreeAddress`
function. By default, LLVM converts them to `LEA64_32r` that requires
64-bit operands, even though these instructions originally use 32-bit
operands. If these values are spilled, an 8-byte slot is needed, like in
the case of `DEC`/`INC`.

Another issue that arises is that having the 32-bit operands requires
`COPY` instructions at the MIR level, which elongates the live intervals
of the value being copied, leading to different interference weights and,
potentially, to different allocation results.

* [layout] Emulate three-address conversion in AArch64

Since we are already emulating the conversion of AArch64's ADD/SUB-rr
instructions to a two-address version, we need to also emulate the conversion
of those back to their original three-address format. This mimics what X86
is doing, by converting two-address additions to a three-address LEA.

Potentially, we need to expand this to other operations, like shifts.