[layout][AArch64] Do not commute operands for sub instructions
When we added the two-address instruction format for AArch64's `add` and
`sub` instructions, we added the capability to commute registers, like
in the case of X86, creating more optimization opportunities. However,
this should only happen for the `add`-like instructions.

X86 differentiates between the two by giving an extra argument to
`ArithBinOp_RF`. We follow the same tactic and check `isSub` in AArch64.