[X86][ISel] Use 64-bit src for MOVSX64rr32 instead of 32-bit (#1)
The equivalent AArch64 instruction that is used to lower
`ISD:SIGN_EXTEND` is `SBFMXri`, which operates on
64-bit operands. This means that it creates a 64-bit subregister and
then copies the original 32-bit register using `INSERT_SUBREG`.

However, after register coalescing, this copy may be eliminated, leaving
only the 64-bit virtual register. If later this register is spilled,
then a 64-bit slot is created, in contrast to X86 which would create a
32-bit slot. Therefore, we need to emulate the same behavior in X86.

We now add the TableGen macro `-DUNIFICO_INSTR_EXTENSION`.