[AArch64][TableGen] Rematerialize LDR+imm instructions
The equivalent X86 instructions, of the form `MOVXXrm`, are
rematerializable, so we choose to do the same with the LDR instructions
that are subclasses of `LoadUI`. Perhaps we will need to do the same in
the future for LDR instructions that are subclasses of `LoadXXRO`,
i.e., which have a register as an offset instead of an immediate.

Currently, this change increases the mean overhead of NPB in AArch64 by
more than 1%, but we will see in the future whether there is a better
alternative (e.g., not rematerialize the MOV-like instructions).

This is still activated using the `UNIFICO_REMAT_RULES` TableGen
directive.

Addresses: https://github.com/systems-nuts/unifico/issues/295