[AArch64] Use the two-address format for floating-point instructions
Covers two-operand and three-operand fp instructions in AArch64. This is needed
for the register allocator, in order to give similar decisions with X86.

We also check the commutability of instructions. We do not commute subtractions
and divisions. Also, for the three-operand instructions, it is still ok if we
treat them as commutable, because like in the case of integer arithmetic the
compiler will only try to commute the first two operands.