[Isel] Assign the correct register class when emitting a COPY
Extends https://github.com/blackgeorge-boom/llvm-project/pull/53/ to
propagate the source register class to the destination not only when
copying from a virtual register, but also when emitting any `COPY`
during isel.

For example, this now covers the case when we lower `Constant:i32<0>`
to `%vr:gpr32temp = COPY $wzr`.