[X86][Isel] Revert inversion of operands for CMOV
Reverts changes in `CMOV` after:
https://github.com/blackgeorge-boom/llvm-unifico/pull/6

We initially though that when lowering `select` to the X86 `CMOV` the
order of operands didn't matter. In fact, it matters, otherwise the wrong
operand is assigned each time the condition is evaluated.

We undo this mistake, and will invert the order of operands in the case
of AArch64 instead, where it doesn't matter, since we have three
operands to play with (dest + two src). We will, however, need to invert
the condition.