Lowering `select` properly for the two architectures (#8)
* [X86][Isel] Revert inversion of operands for CMOV

Reverts changes in `CMOV` after:
https://github.com/blackgeorge-boom/llvm-unifico/pull/6

We initially though that when lowering `select` to the X86 `CMOV` the
order of operands didn't matter. In fact, it matters, otherwise the wrong
operand is assigned each time the condition is evaluated.

We undo this mistake, and will invert the order of operands in the case
of AArch64 instead, where it doesn't matter, since we have three
operands to play with (dest + two src). We will, however, need to invert
the condition.

* [AArch64][Isel] Invert the operands when lowering select to CSEL

In the case of X86, since the equivalent `CMOV` is a two-address instruction, it
needs to invert the operands and the condition of `select`. So, if the
original condition was true, nothing happens, and the destination already has
the correct value since it is constrained to be the first source operand. And if
the original condition was false, then the move happens, and the original
false-value is moved to the destination since the inverted condition evaluates
to true.

We emulate the same behavior, by inverting the operands and the condition when
lowering `select` to `CSEL` in AArch64.

* [AArch64][Isel] Guard the two-address constraint of conditional selects

* [AArch64][TwoAddr] Implement commuting operands for CSELWr/CSELXr

Commuting the operands for conditional selection does not happen
trivially, since if we commute them we need to invert the condition as
well. We emulate the same behavior of X86 and provide the implementation
for commuting `CSEL{W,X}r` instructions.

This will open the way for more optimizations during the
`twoaddressinstruction` pass hopefully, and will also match the X86
behavior during register allocation.