[AArch64] Two-address format for FP instructions and lowering of FABS (#9)
* [AArch64] Use the two-address format for floating-point instructions

Covers two-operand and three-operand fp instructions in AArch64. This is needed
for the register allocator, in order to give similar decisions with X86.

We also check the commutability of instructions. We do not commute subtractions
and divisions. Also, for the three-operand instructions, it is still ok if we
treat them as commutable, because like in the case of integer arithmetic the
compiler will only try to commute the first two operands.

* [AArch64][ISel] Custom-lower fabs/fneg similarly to X86

X86 lowers fabs/fneg of a `f64` value by first converting the value to a
`v2f64` vector and then by using an equivalent logical operator with a
16-byte constant. Then, this operand might be spilled as a 16-byte
value, so we need to do the same custom lowering for AArch64.

Thus, we first convert the `f64` value into a `v2f64`, and then use the
fabs/fneg operator. However, the value seems to be spilled similarly
with 16-bytes.

Activate this by adding `-mattr=+x86-fabs-fneg` for AArch64.

Addresses: https://github.com/systems-nuts/unifico/issues/303