[X86][ISel] Materialize 8-bit zeros differently from 32-bit
The calling conventions of X86 dictate that before each function call
with variadic arguments it should use the `al` register to pass the
number of vector arguments. Because of this, X86 may end up reusing the
zero constant as an arch-specific value and, e.g., pass it to the stack
(it cannot use a CSR because we have forbidden it anyway for `MOV32r0`).
AArch64 on the other hand, uses the zero register in these cases anyway.

One solution to this is to materialize the 8-bit zero constant
differently from others that use `MOV32r0`, and lower it later to the
same 32-bit constant. This way, we "fool" the instruction selection to
avoid using the same zero constant everywhere, by creating a new
constant through `MOV8r0`.

Perhaps in the future we need to revisit this, or extend for 16-bit zero
constants as well.

Used through `UNIFICO_REGALLOC_RULES`.