[layout] Add stack slot alignment capabilities
The `stack-slot-coloring` pass optionally accepts the `-align-stack-slots`
flags. If the flag is activated, then the JSON file
`<module>_stack_slots.txt` is read, where `<module>` is the name of the
source file being compiled (minus `_opt.ll` which is usually the input
format given by Unifico).

The JSON file has the form:

```
{
    "results": {
        "0": 16
    },
    "main": {
        "0": 8
    }
}
```

This means that the module has a function `results` where the frame index
0 has 16-byte alignment. This alignment is used for the frame index
(which will be mapped to a new `Color` as well).

Currently, this JSON is created by parsing offline the
`stack-slot-coloring` pass debug output for every function and frame
index. Then, for both architectures, each slot is examined and the
maximum of their alignments is kept and inserted to the JSON. Finally,
this alignment is used as the size of the stack object.