[regalloc] Remove conditional guards from GPR temp classes
For the AArch64 GPR32temp/GPR64temp classes and for the X86 GR32temp class,
we remove the conditional compilation guards in TableGen, because some of
these classes are referenced in the code (even if they are deactivated through
the Unifico flags). Because not including them in TableGen will make the
compilation fail, we always define these classes in TableGen. Thus, we
can decide whether to actually use them or not, through the respective
llc flags.