[stacktransform] Stack slot copies for arch-specific stack slots
When examining an architecture-specific stack slot, we need to go over
its copies but we only care about store instructions, ignoring loads.
These are that defined the stack slot. If there are multiple, we pick the
closest one to the callsite in the basic block, otherwise if none exists
in the basic block, we give up. Finally, if we manage to find a store,
we repeat the process for the source register of the store, in order to find
its defining instruction(s), as we did for arch-specific virtual register copies.

Also, fixes the identation of some warning messages.

CLoses: https://github.com/blackgeorge-boom/llvm-unifico/issues/19