[stackmaps] Handle arch-specific stack slots like X86 MOVmi (#22)
* Handle memory-to-register operations

Architectures like X86, support operations between a register and the
memory other than loads and stores, e.g., MOV32mi. These often come from
folded loads, like materializing and storing an immediate. To account for these
instructions, we support two ways to define a stack slot that has not
been handled already by the stackmap (architecture-specific):

1. A spill instruction that stores a virtual register to the stack slot.
2. A store instruction that places an immediate in the stack slot.

In the first case, we need to find the definition of the virtual
register being spilled, in order to calculate the machine live
value. In the second case, we can directly calculate the machine
live value from the immediate value. Hence, we extend the check of an
instruction being a store instruction, to include a case where the
value stored is an immediate.

However, for the first case, we need to make sure we avoid cyclic
dependencies where the defining virtual register can be redefined from
the stack slot it had originally defined (e.g., in the case of
spill/refill).

    COPY vreg0, vreg1
    STORE vreg0, <fi#0>
    ...
    STACKMAP 0, 0, <fi#0>
    ...
    LOAD vreg0, <fi#0>

Here, vreg0 is not live because it is redefined after the stackmap.
However, <fi#0> is live, but should be added as an arch-specific because
it was created through a spill. Thus, we track STORE as its definition, 
and, then, we find the definition of vreg0, which comes from vreg1,
constituting a machine-generated value with one operation (Set).

* Add physical register as live-out in sanitizeVregs

When sanitizing a virtual register, if its source is a physical
register not handled in the stackmap, add it as a live-out register.