[stackmaps] Add heuristic for multiple defs of arch-specific value
In `tryToBreakDefMITie`, instead of only trying to find the closest
preceding definition in the same basic block, try to find the first
instruction that satisfies one of the following, if the above fails:

* The instruction is as cheap as a move.
* The instruction is a copy.

This means that it is likely an initialization instruction, such as
a MOV32r0 in X86 or a COPY wzr in AArch64.

If both heuristics fail, we just pick the first instruction for now.

Addresses: https://github.com/blackgeorge-boom/llvm-unifico/issues/23