[stackmaps][X86] Increase LEA coverage
Currently, stackmaps do not handle arch-specific values of the form:

```llvm
  %41:gr64 = LEA64r %stack.0.A, 1, $noreg, 48, $noreg
```

where the value is a `MachineStackObject` with non-zero displacement.

We extend `MachineStackObject` with an optional `Offset` field, so that
we can encode an additional offset from the frame pointer, apart from
the offset of the slot itself.

This is helpful when an arch-specific slot contains a pointer to an array
or a struct element, where the start address of the array or struct is
stored in another slot index. So, the arch-specific value is a slot
index plus an offset.

Closes: https://github.com/blackgeorge-boom/llvm-unifico/issues/25