[stackmaps] Extend sanitizeVregs to return unhandled virtual registers
So far, if sanitizeVregs found an unhandled virtual register in a
MachineInstruction definition, it would flush the value defined by the
instruction and return. Now, instead, we extend it to return the
unhandled virtual register, so that its caller can try instead to find
an arch-specific value for the unhandled register, add it to the
stackmap, and then return to the original instruction.  The caller has
to push the original vreg to a queue, in order to try later to find its
definition and its machine live value.