Make call to sanitizeVregs re-entrant (#30)
* [stackmaps][AArch64] Handle stack object with offset in ADDXri/ORRXri

As we did in: https://github.com/blackgeorge-boom/llvm-unifico/pull/26,
we extend the above instructions to yield `MachineStackObjects` with
potentially extra offset as arch-specific live values.

* [stackmaps] Extend sanitizeVregs to return unhandled virtual registers

So far, if sanitizeVregs found an unhandled virtual register in a
MachineInstruction definition, it would flush the value defined by the
instruction and return. Now, instead, we extend it to return the
unhandled virtual register, so that its caller can try instead to find
an arch-specific value for the unhandled register, add it to the
stackmap, and then return to the original instruction.  The caller has
to push the original vreg to a queue, in order to try later to find its
definition and its machine live value.