[stackmaps] Fix offset when stackmap is pushed downwards
The correct stackmap offset from the beginning of the parent function is
exactly after the corresponding function call. This is supposed to point
where execution will happen in the destination node after migration.

Sometimes the compiler inserts MIR instructions between a function call
and the corresponding stackmap, pushing the stackmap downwards.

```llvm
BL @func1 ...
%14:gpr32temp = LDRWui %stack.0, 0 ...
ADJCALLSTACKUP 0, 0, implicit-def $sp ...
ADJCALLSTACKDOWN 0, 0, implicit-def $sp ...
PCN_STACKMAP 3, 0, %14:gpr32temp ...
```

In this case, to have the correct stackmap offset, we keep track of the
next instruction after the call through a temporary symbol. Then, the
correct stackmap offset from the parent function becomes the offset of
this symbol, and not the offset of the stackmap itself.