Revert "[layout] Add SHF_MERGE flag for section with known entry size."
This reverts commit a4869d678479aa9c38a82de94cffe5e94ea7718c.

For some reason, this change messes up with the
`.stack_transform.unwind` locations, so after running `make
post_process`, there aren't enough records, compared to what the
`.stack_transform.unwind_arange` has recorded. This breaks the stack
transformation code, e.g., `recode` in the TransProc project, for
comparison with the baseline.

It is useful to have this change in the `development` branch, because we
often want to experiment with stack transformation in unifico-compiled
binaries (even though they normally wouldn't use transfromation
metadata) for comparison with related work.

This was introduced as one of the commits to fix:
https://github.com/systems-nuts/unifico/issues/91