[stackmaps] Make IR stackmap search more robust
When lowering the stackmaps, the MachineInstruction representing the
stackmap needs to find the corresponding IR instruction, to make sure
that all of the operands have been lowered properly.

However, optimizations like those introduced by `BranchFolding.cpp` can
introduce new blocks and move instructions around, breaking the IR block
to MIR block mapping between instructions. The default case, is to find
the IR block from the MIR and iterate until the IR stackmap is found. We
introduce two additional searches:

- 1. If the IR instruction was not found in the block, try to search the
successor IR block as well, to cover for the case where the MIR
instruction was hoisted in the predecessor.
- 2. If the IR Basic Block was not found in the first place, like in the
case of an optimization that split the original IR basic block and the
stackmap MIR ended up in the "fake" block without a corresponding IR
block, search across the entire function for the IR stackmap.